services/enqueue-api/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
google-cloud-pubsub==2.18.4
google-cloud-logging==3.8.0
google-cloud-run==0.10.5
httpx==0.27.0
